+++ tests/mat3.rs
@@ -277,6 +277,16 @@ macro_rules! impl_mat3_tests {
             assert_eq!(format!("{}", a), "[[1, 2, 3], [4, 5, 6], [7, 8, 9]]");
         }
 
+        #[test]
+        fn test_mat3_to_from_slice() {
+            const MATRIX1D: [$t; 9] = [1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0];
+            let m = $mat3::from_cols_slice(&MATRIX1D);
+            assert_eq!($mat3::from_cols_array(&MATRIX1D), m);
+            let mut out: [$t; 9] = Default::default();
+            m.write_cols_to_slice(&mut out);
+            assert_eq!(MATRIX1D, out);
+        }
+
         #[cfg(feature = "std")]
         #[test]
         fn test_sum() {