+++ tests/vec3.rs
@@ -416,8 +416,8 @@ macro_rules! impl_vec3_tests {
         fn test_to_from_slice() {
             let v = $vec3::new(1 as $t, 2 as $t, 3 as $t);
             let mut a = [0 as $t, 0 as $t, 0 as $t];
-            v.write_to_slice_unaligned(&mut a);
-            assert_eq!(v, $vec3::from_slice_unaligned(&a));
+            v.write_to_slice(&mut a);
+            assert_eq!(v, $vec3::from_slice(&a));
         }
 
         #[cfg(feature = "std")]